+++ cypress/integration/wipe.spec.js
@@ -14,7 +14,7 @@ describe('Wipe', () => {
   testSetup()
 
   it('Deleting checkout history', () => {
-    cy.getVtexItems().then(vtex => {
+    cy.getVtexItems().then((vtex) => {
       cy.request({
         url: `https://productusqa.vtexcommercestable.com.br/api/dataentities/checkoutcustom/search`,
         qs: {
@@ -27,7 +27,7 @@ describe('Wipe', () => {
           'REST-Range': 'resources=0-50',
         },
         ...FAIL_ON_STATUS_CODE,
-      }).then(response => {
+      }).then((response) => {
         expect(response.status).to.equal(200)
         for (const { id } of response.body) {
           cy.deleteDocumentInMasterData('checkoutcustom', id)
@@ -36,15 +36,18 @@ describe('Wipe', () => {
     })
   })
   it('Getting user & then deleting addresses associated with that user', () => {
-    cy.getVtexItems().then(vtex => {
-      cy.searchInMasterData(ENTITIES.CLIENTS, vtex.robotMail).then(clients => {
-        cy.searchInMasterData(ENTITIES.ADDRESSES, clients[0].id).then(
-          addresses => {
-            for (const { id } of addresses)
-              cy.deleteDocumentInMasterData(ENTITIES.ADDRESSES, id)
-          }
-        )
-      })
+    cy.getVtexItems().then((vtex) => {
+      cy.searchInMasterData(ENTITIES.CLIENTS, vtex.robotMail).then(
+        (clients) => {
+          cy.searchInMasterData(ENTITIES.ADDRESSES, clients[0].id).then(
+            (addresses) => {
+              for (const { id } of addresses) {
+                cy.deleteDocumentInMasterData(ENTITIES.ADDRESSES, id)
+              }
+            }
+          )
+        }
+      )
     })
   })
 })