+++ cypress/support/commands.js
@@ -5,7 +5,7 @@ const checkOutJson = '.checkout.json'
 
 // Set checkout items
 Cypress.Commands.add('setCheckOutItem', (checkItem, checkvalue) => {
-  cy.readFile(checkOutJson).then(items => {
+  cy.readFile(checkOutJson).then((items) => {
     items[checkItem] = checkvalue
     cy.writeFile(checkOutJson, items)
   })
@@ -13,7 +13,7 @@ Cypress.Commands.add('setCheckOutItem', (checkItem, checkvalue) => {
 
 // Get checkout items
 Cypress.Commands.add('getCheckOutItems', () => {
-  cy.readFile(checkOutJson).then(items => {
+  cy.readFile(checkOutJson).then((items) => {
     return items
   })
 })
@@ -32,18 +32,14 @@ Cypress.Commands.add('openStoreFront', (login = false) => {
 Cypress.Commands.add('gotoProductDetailPage', () => {
   cy.get(selectors.ProductAnchorElement)
     .should('have.attr', 'href')
-    .then(href => {
-      cy.get(generateAddtoCartCardSelector(href))
-        .first()
-        .click()
+    .then((href) => {
+      cy.get(generateAddtoCartCardSelector(href)).first().click()
     })
 })
 
 Cypress.Commands.add('openProduct', (product, detailPage = false) => {
   // Search product in search bar
-  cy.get(selectors.Search)
-    .should('be.not.disabled')
-    .should('be.visible')
+  cy.get(selectors.Search).should('be.not.disabled').should('be.visible')
 
   cy.get(selectors.Search).type(`${product}{enter}`)
   // Page should load successfully now Filter should be visible
@@ -57,18 +53,14 @@ Cypress.Commands.add('openProduct', (product, detailPage = false) => {
   }
 })
 
-Cypress.Commands.add('removeProduct', product => {
-  cy.get(`#item-remove-${product}`)
-    .should('be.visible')
-    .click()
+Cypress.Commands.add('removeProduct', (product) => {
+  cy.get(`#item-remove-${product}`).should('be.visible').click()
   cy.contains('Your cart is empty')
 })
 
 Cypress.Commands.add('checkoutProduct', () => {
   cy.get(selectors.TotalPrice).should('be.visible')
-  cy.get(selectors.ProceedtoCheckout)
-    .should('be.visible')
-    .click()
+  cy.get(selectors.ProceedtoCheckout).should('be.visible').click()
   cy.get(selectors.CartTimeline, { timeout: 30000 })
     .should('be.visible')
     .click({ force: true })