+++ node/graphql/index.ts
@@ -80,7 +80,7 @@ const schema = {
 
 const replacer = (template: string, keys: any) => {
   return template.replace(/"?{{\w+}}"?/g, (key: string) => {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore: Object is possibly 'null'.
     return keys[key.match(/\w+/)[0]] || 'false'
   })